+++ EssentialFeed/EssentialFeed/FeedAPI/RemoteFeedItem.swift
@@ -1,0 +1,16 @@
+//
+//  RemoteFeedItem.swift
+//  EssentialFeed
+//
+//  Created by Aaron Huánuco on 13/09/2020.
+//  Copyright © 2020 Aaron Huánuco. All rights reserved.
+//
+
+import Foundation
+
+internal struct RemoteFeedItem: Decodable {
+    internal let id: UUID
+    internal let description: String?
+    internal let location: String?
+    internal let image: URL
+}
